シンプルなexampleのリファクタ場所変更
